Code refactoring to stick to Gepetto package style.
