+++ CMakeLists.txt
@@ -1,25 +1,21 @@
 CMAKE_MINIMUM_REQUIRED(VERSION 2.6)
 
-# INCLUDE(cmake/base.cmake)
-# INCLUDE(cmake/eigen.cmake)
+INCLUDE(cmake/base.cmake)
+INCLUDE(cmake/eigen.cmake)
 
 SET(CXX_DISABLE_WERROR True)
-# SET(CMAKE_VERBOSE_MAKEFILE True)
+SET(CMAKE_VERBOSE_MAKEFILE True)
 
-SET(PROJECT_NAME DDPsolver)
+SET(PROJECT_NAME ddp-actuator-solver)
 SET(PROJECT_DESCRIPTION "DDP/iLQR solver for robotics actuators command")
-SET(PROJECT_URL "")
+SET(PROJECT_URL "https://github.com/stack-of-tasks/ddp-actuator-solver")
 
-# ADD_REQUIRED_DEPENDENCY("eigen3 >= 3.0.5")
+ADD_REQUIRED_DEPENDENCY("eigen3 >= 3.0.5")
+ADD_REQUIRED_DEPENDENCY("qpOASES")
 
-# SETUP_PROJECT()
+INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/include)
 
-LIST(APPEND CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_SOURCE_DIR}/private_cmake)
-MESSAGE(STATUS "CMAKE_MODULE_PATH: ${CMAKE_MODULE_PATH}")
-FIND_PACKAGE(qpOASES)
-SEARCH_FOR_QPOASES()
+ADD_SUBDIRECTORY(src)
+ADD_SUBDIRECTORY(test)
 
-
-SET(LIBRARY_OUTPUT_PATH ${PROJECT_NAME}/lib)
-
-ADD_SUBDIRECTORY(cpp)
+SETUP_PROJECT()
