--- cpp/src/CMakeLists.txt
@@ -1,1 +1,0 @@
-
