+++ include/ddp-actuator-solver/costfunctionromeoactuator.h
